+++ 4_cv_basics/6_morphology/include/morphology.hpp
